🌟 Webhook support.

+++ src/modules/NSFWProcessor.py
@@ -3,6 +3,7 @@ from time import time
 from discord import Message, File
 from nudenet import NudeClassifier, NudeDetector
 from utilsx.discord import Cog
+from utilsx.discord.objects import Footer
 
 from run import Bot
 from src.utils.Logger import Logger
@@ -19,6 +20,9 @@ class NSFWProcessor(Cog):
 
     @Cog.listener()
     async def on_message(self, message: Message):
+        if message.author.bot or message.channel.is_nsfw():
+            return
+
         Logger.debug(f"Started processing {message.id}")
         new_images = []
         if message.attachments:
@@ -38,11 +42,26 @@ class NSFWProcessor(Cog):
                     detector.censor(f"./tmp/{attachment.filename}", name)
                     new_images.append(File(name, attachment.filename))
 
-        for image in new_images:
-            em = await self.embed(message.channel, f"Auto detected NSFW in message from {message.author.mention}!",
-                                  get_embed=True, image=f"attachment://{image.filename}")
-            await self.send(message.channel, "", embed=em, file=image)
-            await message.delete()
+        if new_images:
+            hooks = await message.channel.webhooks()
+            sender = None
+            for hook in hooks:
+                if hook.user == self.bot.user:
+                    sender = hook
+                    break
+
+            if sender is None:
+                sender = await message.channel.create_webhook(name="Guardian",
+                                                              reason="No existing webhook was found, creating one!")
+
+            for image in new_images:
+                em = await self.embed(sender, "", get_embed=True, image=f"attachment://{image.filename}",
+                                      footer=Footer("Auto detected NSFW, removed by Guardian",
+                                                    self.bot.user.avatar_url))
+                # await self.send(sender, "", embed=em, file=image)
+                await sender.send(content=message.content, username=str(message.author),
+                                  avatar_url=message.author.avatar_url, file=image, embed=em)
+                await message.delete()
 
         Logger.debug(f"Successfully processed {message.id}")
